requirements.txt
flask>=3.1.0
flask-session>=0.8.0
groq>=0.13.0
boto3>=1.35.0
bcrypt>=4.2.0
python-dotenv>=1.0.1
gunicorn>=22.0.0
